redo f16 checks
