+++ utils/bazel/llvm-project-overlay/libc/BUILD.bazel
@@ -2062,11 +2062,10 @@ libc_support_library(
     name = "expxf16",
     hdrs = ["src/math/generic/expxf16.h"],
     deps = [
-        ":__support_cpp_array",
         ":__support_fputil_cast",
         ":__support_fputil_fp_bits",
         ":__support_fputil_nearest_integer",
-        ":__support_fputil_polyeval",
+        ":__support_math_expf16_utils",
     ],
 )
 
@@ -2097,6 +2096,18 @@ libc_support_library(
 )
 
 libc_support_library(
+    name = "__support_math_expf16_utils",
+    hdrs = ["src/__support/math/expf16_utils.h"],
+    deps = [
+        ":__support_cpp_array",
+        ":__support_fputil_nearest_integer",
+        ":__support_fputil_polyeval",
+        ":__support_macros_attributes",
+        ":llvm_libc_macros_float16_macros"
+    ],
+)
+
+libc_support_library(
     name = "__support_math_expf16",
     hdrs = ["src/__support/math/expf16.h"],
     deps = [
@@ -2112,6 +2123,8 @@ libc_support_library(
         ":__support_fputil_polyeval",
         ":__support_fputil_rounding_mode",
         ":__support_macros_optimization",
+        ":__support_math_expf16_utils",
+        ":llvm_libc_macros_float16_macros"
     ],
 )
